+++ .github/workflows/ci.yaml
@@ -38,14 +38,12 @@ jobs:
         with:
           go-version: 1.14
       - name: build
-        run: |
-          go build
+        run: make
         env:
           GOOS: ${{ matrix.os }}
           GOARCH: ${{ matrix.arch }}
       - name: test
         if: matrix.os == 'linux' && matrix.arch == 'amd64'
         run: |
-          set -e
-          ./nebula-console2.0 -e 'exit'
-          ./nebula-console2.0 -f demo.nGQL
+          ./nebula-console -e 'exit'
+          ./nebula-console -f demo.nGQL
