+++ nebula/table.go
@@ -4,7 +4,7 @@
  * attached with Common Clause Condition 1.0, found in the LICENSES directory.
  */
 
-package main
+package nebula
 
 import (
 	"bytes"
